// src/payload-types.ts
export interface Repair {
  id: number;
  name: string;
  price: number;
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}